requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
gunicorn==22.0.0
# MongoDB
mongoengine==0.29.3
pymongo==4.16.0
bcrypt==5.0.0
# JWT
PyJWT==2.12.1

# ── System dependencies (not pip-installable) ─────────────────────────────────
# terraform >= 1.3.0   https://developer.hashicorp.com/terraform/install
#   Used by apps/terraform/runner.py to execute infrastructure deployments.
#   install.sh installs it automatically; Dockerfile bundles it at build time.
